Refresh project summary table after loading data
Refactor - remove dependency logic -> GUI
